Adds .iml file to gitignore, removes .iml file from repo
The pom file is the only project configuration file that's
really needed at this point. The iml file keeps causing git
to recognize modifications to it. It's an automatically
generated file so the classical git wisdom says to just leave
it out of the repo. Everybody using IntelliJ will have a new
one generated for them the next time they open the project -
we can just keep our respective .iml's to ourselves ;-)